+++ packages/sphinx-codelinks/tests/test_src_trace.py
@@ -59,7 +59,7 @@ from sphinx_codelinks.sphinx_extension.source_tracing import set_config_to_sphin
             [
                 "Project 'dcdc' has the following errors:",
                 "Schema validation error in field 'exclude': 123 is not of type 'string'",
-                "Schema validation error in field 'comment_type': 'java' is not one of ['cpp', 'cs', 'python', 'rust', 'yaml']",
+                "Schema validation error in field 'comment_type': 'java' is not one of ['cpp', 'cs', 'go', 'python', 'rust', 'yaml']",
                 "Schema validation error in field 'gitignore': '_true' is not of type 'boolean'",
                 "Schema validation error in field 'include': 345 is not of type 'string'",
                 "Schema validation error in field 'src_dir': ['../dcdc'] is not of type 'string'",
@@ -192,6 +192,10 @@ def test_src_tracing_config_positive(make_app: Callable[..., SphinxTestApp], tmp
         (
             Path("doc_test") / "id_required",
             Path("doc_test") / "id_required",
+        ),
+        (
+            Path("doc_test") / "go_basic",
+            Path("doc_test") / "go_basic",
         ),
     ],
 )